refactored namespacing across edited files, added instantiation of MySqlConnection to replace new PDO line 14 login.php
