+++ src/public/login.php
@@ -3,7 +3,7 @@
 require_once '../../vendor/autoload.php';
 try {
     if(isset($_POST['email'])){
-    $email = new \phbees\pokedex\Email($_POST['email']);
+    $email = new Pokedex\Classes\Email($_POST['email']);
     } else {
         header('Location:loginPage.php?invalid_email=1');
     }
@@ -11,7 +11,7 @@ try {
     header('Location:loginPage.php?invalid_email=1');
 }
 
-$db = new PDO('mysql:host=127.0.0.1;dbname=Pokedex','root');
-$users = new \phbees\pokedex\Users($db,$email);
+$db = new Pokedex\Classes\MySqlConnection();
+$users = new Pokedex\Classes\Users($db,$email);
 $users->grabIdFromDb();
 $users->checkIfEnteredEmailExists();
